+++ Sampler/Public/Get-CodeCoverageThreshold.ps1
@@ -36,20 +36,29 @@ function Get-CodeCoverageThreshold
     # If no codeCoverageThreshold configured at runtime, look for BuildInfo settings.
     if ([String]::IsNullOrEmpty($RuntimeCodeCoverageThreshold))
     {
-        if ($BuildInfo.ContainsKey('Pester') -and $BuildInfo.Pester.ContainsKey('CodeCoverageThreshold'))
+        if ($BuildInfo -and $BuildInfo.ContainsKey('Pester') -and $BuildInfo.Pester.ContainsKey('CodeCoverageThreshold'))
         {
             $codeCoverageThreshold = $BuildInfo.Pester.CodeCoverageThreshold
+
             Write-Debug -Message "Loaded Code Coverage Threshold from Config file: $codeCoverageThreshold %."
+        }
+        elseif ($BuildInfo -and $BuildInfo.ContainsKey('Pester') -and $BuildInfo.Pester.ContainsKey('Configuration') -and $BuildInfo.Pester.Configuration.CodeCoverage.CoveragePercentTarget)
+        {
+            $codeCoverageThreshold = $BuildInfo.Pester.Configuration.CodeCoverage.CoveragePercentTarget
+
+            Write-Debug -Message "Loaded Code Coverage Threshold from Config file in Pester advanced configuration: $codeCoverageThreshold %."
         }
         else
         {
             $codeCoverageThreshold = 0
+
             Write-Debug -Message "No code coverage threshold value found (param nor config), using the default value."
         }
     }
     else
     {
         $codeCoverageThreshold = [int] $RuntimeCodeCoverageThreshold
+
         Write-Debug -Message "Loading CodeCoverage Threshold from Parameter ($codeCoverageThreshold %)."
     }